covid_data_pipeline/requirements.txt
pyspark==3.5.1
matplotlib
pandas
psycopg2-binary